FIX: client file managment
