+++ src/cyhole/core/__init__.py
@@ -1,4 +1,5 @@
-from ..core.api import APICaller, APIClient, AsyncAPIClient
+from ..core.api import APICaller
+from ..core.client import APIClient, AsyncAPIClient
 
 __all__ = [
     "APICaller",
