+++ tests/requirements.txt
@@ -1,4 +1,5 @@
 pytest>=8.2.0
 pytest-mock>=3.14.0
+pytest-asyncio>=0.23.7
 coverage>=7.5.3
 ruff>=0.4.6
